[이성훈] 회사 유저 장바구니 상품 리스트 조회 및 상세 조회 구현

+++ src/main/java/com/springboot/bizconnect/entity/Inquiry.java
@@ -30,7 +30,7 @@ public class Inquiry extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "inquiry_no", nullable = false)
-    private Integer inquiryNo;
+    private Long inquiryNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_no", nullable = false)
